refactor(agents): reorganize host files under agents/, split catalog into agents/catalog/
Two compounding reorganizations, landed together because both stem
from this subsystem outgrowing a flat layout:

1. ADR 0042 phase 7 (trigger widened from "quirks only" to plain
   navigability): claude/codex/cursor/copilot moved out of
   agent-catalog.ts into agents/catalog/<id>.ts as
   build<Id>AgentDefinition(deps) factories (deps: { marker,
   buildHookCommand }, now a shared HookAgentDeps type); grok/opencode
   moved as plain exported objects. agent-catalog.ts drops from 986 to
   ~670 lines.

2. extension-host/ had grown to ~187 entries. Every agent-*/agents-*
   file that lived directly under it now lives under a new agents/
   folder instead; the pre-existing per-agent-definition folder
   (agents/pi.ts) was renamed to agents/catalog/ first, freeing the
   agents name. Filenames were not shortened to drop their
   agent-/agents- prefix, to avoid touching every doc/ADR/RFC that
   names these files today — a directory move TypeScript resolves
   transparently is a much smaller, safer change.

Pure move + import-path fixes (including vi.mock() specifiers, easy
to miss since they're string literals, not import statements) — no
behavior change. docs/adding-a-coding-agent.md, ADR 0042 (new
phase-7 implementation note), and ADR 0043's stale forward-reference
updated in the same change.

Stacks on #437 (agent-title-identity-prefix) — this branch's
agent-catalog.ts already carries that PR's titleIdentityPrefix/
stripAgentTitleIdentityPrefix, just relocated.
