+++ packages/extension-host/src/extension-host/agents/agent-activity-model.test.ts
