+++ packages/extension-host/src/extension-host/agents/agent-catalog.test.ts
