+++ packages/extension-host/src/extension-host/agents/agent-detection-dispatch.ts
