+++ packages/extension-host/src/extension-host/agents/agent-hook-events.test.ts
