+++ packages/extension-host/src/extension-host/agents/agent-hook-script.ts
