+++ packages/extension-host/src/extension-host/agents/agent-osc-detectors.test.ts
