+++ packages/extension-host/src/extension-host/agents/agent-resume-hint.ts
