+++ packages/extension-host/src/extension-host/agents/agents-channel.ts
@@ -1,4 +1,4 @@
-import { createHostChannel } from "./output-store";
+import { createHostChannel } from "../output-store";
 
 /** Output-panel channel for `ctx.agents` diagnostics — visible in Silo's own
  * Output window, not the devtools console. */
