+++ packages/extension-host/src/extension-host/agents/catalog/pi.test.ts
