+++ packages/extension-host/src/extension-host/context.ts
@@ -40,7 +40,7 @@ import { getEditorService } from "./editor-service";
 import { getLayoutService } from "./layout-service";
 import { getScopedProcessService } from "./process-service";
 import { getScopedProcessesService } from "./processes-service";
-import { getAgentsService } from "./agents-service";
+import { getAgentsService } from "./agents/agents-service";
 import { getTerminalService } from "./terminal-service";
 import { getScopedFileService } from "./file-service";
 import { getScopedSearchService } from "./search-service";
